refactor: service의 money확인 후 lotto생성 책임 LottoShop으로 이동, 테스트 리팩토링
